Push test notifications to iOS sandbox
This is used in deciding whether to use your iOS Sandbox or Production push certificate when sending a push when both have been uploaded. Set to the iOS provisioning profile that was used to build your app. 1 = Development, 2 = Ad-Hoc. Omit this field for App Store builds. https://documentation.onesignal.com/docs/players-add-a-device